validator and test validator

+++ apolloServer/src/main/java/com/apollo/service/StrategyServiceImpl.java
@@ -2,6 +2,7 @@ package com.apollo.service;
 
 import com.apollo.entities.Strategy;
 import com.apollo.repos.StrategyRepository;
+import com.apollo.util.Validator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -10,6 +11,8 @@ import java.util.Collection;
 
 @Service
 public class StrategyServiceImpl implements StrategyService {
+    @Autowired
+    Validator validator;
     @Autowired
     private StrategyRepository dao;
 
@@ -32,32 +35,38 @@ public class StrategyServiceImpl implements StrategyService {
 
     @Override
     public String createOrUpdate(Strategy s) {
-        if(s.getId() == null) {
+        if (s.getId() == null) {
             s.setId(0);
         }
+        if (validator.validateStrategy(s)) {
+            if (dao.exists(s.getId())) {
+                dao.save(s);
+                return "Updated strategy";
 
-        if (dao.exists(s.getId())) {
-            dao.save(s);
-            return "Updated strategy";
-
+            } else {
+                return dao.save(s).getId().toString();
+            }
         } else {
-            return dao.save(s).getId().toString();
+            return "Strategy parameters were invalid";
         }
     }
 
     // at the /strategy/{id} endpoint
     @Override
     public Strategy getById(int id) {
-        return dao.findOne(id);
+        if (validator.validateId(id)) {
+            return dao.findOne(id);
+        } else
+            return null;
     }
 
     @Override
     public String deleteById(int id) {
-        if (dao.exists(id)) {
+        if (dao.exists(id) && validator.validateId(id)) {
             dao.delete(id);
             return "Deleted Strategy with ID number " + id;
         } else
-            return "Unable to find Strategy with that ID";
+            return "Unable to find Strategy with that ID or ID invalid";
     }
 
     // at the /strategy/start endpoint
